+++ ui/components/path/index.js
@@ -40,6 +40,8 @@ Path.defaultProps = {
  * @param {*} props
  * @prop {boolean} hasCaoching - When true, adds the Coaching Button Toggle and preps Path Track for coaching support
  * @prop {boolean} coachingOpen - When true, displays the Coaching Button Toggle in an active state
+ * @prop {boolean} isSmallRegion - When true, the Mark as Complete button will be full width
+ * @prop {boolean} isMobile - When true, removes the coaching toggle and the scroll buttons
  * @prop {boolean} hasOverflow - When true, presents the Path Steps in a scroller wrapper with control buttons
  * @prop {string} coachingId - Id of Coaching container element
  * @prop {string} stageName - Visible name of the current stage
@@ -52,7 +54,9 @@ export const PathTrack = props => {
     hasOverflow,
     coachingId,
     stageName,
-    actionButtonLabel
+    actionButtonLabel,
+    isSmallRegion,
+    isMobile
   } = props;
 
   return (
@@ -86,7 +90,7 @@ export const PathTrack = props => {
               {props.children}
             </ul>
 
-            {hasOverflow ? (
+            {hasOverflow && !isMobile ? (
               <div className="slds-path__scroll-controls">
                 <ButtonIcon
                   symbol="left"
@@ -111,9 +115,24 @@ export const PathTrack = props => {
 
       <div className="slds-grid slds-path__action">
         <span className="slds-path__stage-name">Stage: {stageName}</span>
-        <Button isBrand className="slds-path__mark-complete">
+        <Button
+          isBrand
+          className={classNames(isSmallRegion && 'slds-button_stretch')}
+        >
           {actionButtonLabel}
         </Button>
+        {isSmallRegion && (
+          <div className="slds-path__trigger-region-small">
+            <Button
+              isNeutral
+              className={classNames(isSmallRegion && 'slds-button_stretch')}
+              aria-expanded={coachingOpen}
+              aria-controls={coachingId}
+            >
+              {coachingOpen ? 'Show Less' : 'Show More'}
+            </Button>
+          </div>
+        )}
       </div>
     </div>
   );
@@ -125,7 +144,9 @@ PathTrack.propTypes = {
   hasOverflow: PropTypes.bool,
   coachingId: PropTypes.string,
   stageName: PropTypes.string,
-  actionButtonLabel: PropTypes.node
+  actionButtonLabel: PropTypes.node,
+  isSmallRegion: PropTypes.bool,
+  isMobile: PropTypes.bool
 };
 
 PathTrack.defaultProps = {
@@ -263,7 +284,6 @@ export const PathCoaching = props => {
             direction="stacked"
             snapshot={coachingData}
             isViewMode
-            hasInlineEdit
           />
         </div>